requirements.txt
aiohttp==2.3.2
async-timeout==2.0.0
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
idna==2.6
multidict==3.3.2
python-engineio==1.7.0
python-socketio==1.8.3
requests==2.18.4
six==1.11.0
urllib3==1.22
yarl==0.14.2
